fix(CodeEditor): Prevent loss of styles/JS
The Monaco code editor injects styles and JavaScript tags into the header. Usually, in React, this is fine. In Astro, with transitions, this breaks down. I tried manually adding these and it broke Monaco. Instead, forcing a reload and trying to fix dark mode seemed like a better way to go.